Make fix of index.js to not try to drop tables and reconect when testing is running

+++ back-end/fileParseApi/index.js
@@ -1,36 +1,36 @@
 const express = require("express");
 const cors = require('cors');
 const fileParserRoute = require('./routes/fileParseRoute');
+const fileUpload = require('express-fileupload');
 const bodyParser = require('body-parser');
 const app = express();
-
-const db = require("./models/index");
-db.sequelize.sync({ force: true }).then(() => {
-    console.log("Drop and re-sync db.");
-  });
 
 const corsOptions = {
     origin: '*',
     optionSucessStatus: 200
 };
 
+app.use(fileUpload());
 app.use(cors(corsOptions));
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: true }));
 
 app.use(express.static("files"));
 
-/* app.use(express.json({
-    limit: "20mb"
-}));
+const db = require("./models/index");
 
-app.use(express.urlencoded({
-    limit: "20mb",
-    extended: true
-})); */
 
 app.use('/api', fileParserRoute);
 
-const port = process.env.PORT || 3003;
 
-module.exports = app.listen(port, () =>  console.log(`Listening on port ${port}....`));
+
+const port = process.env.PORT || 3001;
+if (require.main === module) {
+    //inicia o servidor
+    app.listen(port, () =>  console.log(`Listening on port ${port}....`));
+    db.sequelize.sync({ force: true }).then(() => {
+        console.log("Drop and re-sync db.");
+      });
+}
+
+module.exports = app
